Now mysql port is a int32, to be consistent with other places. Also fixing all linter issues.

+++ go/cmd/mysqlctl/mysqlctl.go
@@ -167,7 +167,7 @@ func main() {
 	flag.Parse()
 
 	tabletAddr = netutil.JoinHostPort("localhost", int32(*port))
-	mycnf := mysqlctl.NewMycnf(uint32(*tabletUID), *mysqlPort)
+	mycnf := mysqlctl.NewMycnf(uint32(*tabletUID), int32(*mysqlPort))
 
 	if *mysqlSocket != "" {
 		mycnf.SocketFile = *mysqlSocket
